+++ logprep/processor/domain_resolver/processor.py
@@ -178,7 +178,6 @@ class DomainResolver(Processor):
                         list_path.touch()
                         list_path.write_bytes(GetterFactory.from_string(tld_list).get_raw())
                 downloaded_tld_lists_paths.append(f"file://{str(list_path.absolute())}")
-            self._config.tld_lists = downloaded_tld_lists_paths
             logger.debug("finished tldlists download...")
 
     def _apply_rules(self, event, rule):