+++ tests/unit/processor/domain_label_extractor/test_domain_label_extractor.py
@@ -1,6 +1,7 @@
 # pylint: disable=protected-access
 # pylint: disable=missing-docstring
 
+import copy
 import hashlib
 import os
 import shutil
@@ -347,7 +348,9 @@ class TestDomainLabelExtractor(BaseProcessorTestCase):
         tld_list_content = tld_list_path.read_bytes()
         expected_checksum = hashlib.md5(tld_list_content).hexdigest()  # nosemgrep
         responses.add(responses.GET, tld_list, tld_list_content)
-        self.object._config.tld_lists = [tld_list]
+        config = copy.deepcopy(self.CONFIG)
+        config["tld_lists"] = [tld_list]
+        self.object = Factory.create({"domain_label_extractor": config})
         self.object.setup()
         logprep_tmp_dir = Path(tempfile.gettempdir()) / "logprep"
         downloaded_file = logprep_tmp_dir / f"{self.object.name}-tldlist-0.dat"
@@ -370,7 +373,9 @@ class TestDomainLabelExtractor(BaseProcessorTestCase):
         pre_existing_content = "file exists already"
         tld_temp_file.touch()
         tld_temp_file.write_bytes(pre_existing_content.encode("utf8"))
-        self.object._config.tld_lists = [tld_list]
+        config = copy.deepcopy(self.CONFIG)
+        config["tld_lists"] = [tld_list]
+        self.object = Factory.create({"domain_label_extractor": config})
         self.object.setup()
         assert tld_temp_file.exists()
         assert tld_temp_file.read_bytes().decode("utf8") == pre_existing_content